Extract node metrics into module
We extract the metrics registration code from the `service` module and
put it into its own module. This should make the substrate update that
changes the metrics APIs easier.
